Add archive option
This adds the option to create named archives.

It can be desireable to have specific versions of the code exist in
static locations to be referenced in the future, or to be used as a
stable version in other projects.

By calling the script with the -a option, the files in /stable will
be copied to a new folder in the /versions directory, named with the
current date in the YYYYMMDD format.