Cleanup: Make output cleaner
This cleans up the script output.

It adds a "success" function to supplement the "say_done" function
to provide more detailed success status messages.

It adds a "title" function to provide a more impactful way to share
the title of an action than the standard "debug" messages.

It also cleans up the new lines.