Make newlines at the end of config files optional
The prior version of this script used a simple "while read" "do" format
for reading the config files. This worked, but required a newline to
be at the end of each file, otherwise hitting the EOF caused the script
to ignore the last line.

This changes the loops to test a flag instead of the read exit status,
allowing them to read all the lines, regardless of if a newline exists
or not.