Make .trt/ignores lines relative each repo's output destination
Prior to this commit, lines in the .trt/ignores file would only exclude
a file if the path matched the path in the source location.

For example, if project1 and project2 both had a foo/bar.sh file, and
the .trt/ignores file had a line matching "foo/bar.sh", the file
would be blocked from both.

However, if one of the repos had their output redirected in the .trt/repos
file (ex: /home/project1/stable/ > app), it wasn't possible to only specify
to block the app/foo/bar.sh one. This is because the blocking was only relative
to the source.

This change modifies the tool to exclude files that match the output destination.
Ignores can be made relative to their location in the current project, not
the source.